# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface QuestionPreview {
  id: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  subtopic?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  questionType?: string | null;
  /** @nullable */
  countryScope?: string | null;
  question: string;
  /** @nullable */
  expectedAnswerShort?: string | null;
}

export interface ContactInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
  /** @maxLength 200 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  message: string;
}

export interface ContactResult {
  ok: boolean;
}

export interface Provider {
  id: string;
  defaultModel: string;
  available: boolean;
  requiresKey: boolean;
}

export interface BenchmarkConfig {
  providers: Provider[];
  topics: string[];
  difficulties: string[];
  questionTypes: string[];
  judgeModel: string;
  judgeAvailable: boolean;
  totalQuestions: number;
}

export interface RunInput {
  /** @minItems 1 */
  models: string[];
  /** @nullable */
  limit?: number | null;
  /** @nullable */
  topic?: string | null;
  dryRun?: boolean;
  noEval?: boolean;
}

export interface Run {
  id: string;
  /** queued, running, completed, failed or interrupted */
  status: string;
  /** empty, query, evaluate, report or done */
  phase: string;
  models: string[];
  /** @nullable */
  limit?: number | null;
  /** @nullable */
  topic?: string | null;
  dryRun: boolean;
  noEval: boolean;
  completed: number;
  total: number;
  /** @nullable */
  questionCount?: number | null;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  judgeModel?: string | null;
}

export interface ModelSummary {
  provider: string;
  model: string;
  /**
     * Approximate model size class (e.g. Petit, Moyen, Grand)
     * @nullable
     */
  size?: string | null;
  nQuestions: number;
  nErrors: number;
  /** @nullable */
  avgLatency?: number | null;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  uncertaintyHandling?: number | null;
  /** @nullable */
  justificationQuality?: number | null;
  /** @nullable */
  sourceAwareness?: number | null;
  /** @nullable */
  regulatoryHallucinationRisk?: number | null;
  /** @nullable */
  overallScore?: number | null;
}

export interface GroupSummary {
  /** topic or difficulty label */
  group: string;
  provider: string;
  model: string;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  uncertaintyHandling?: number | null;
  /** @nullable */
  justificationQuality?: number | null;
  /** @nullable */
  sourceAwareness?: number | null;
  /** @nullable */
  regulatoryHallucinationRisk?: number | null;
}

export interface ResultRow {
  questionId: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  subtopic?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  questionType?: string | null;
  /** @nullable */
  countryScope?: string | null;
  /** @nullable */
  question?: string | null;
  provider: string;
  model: string;
  /** @nullable */
  rawResponse?: string | null;
  /** @nullable */
  latencySeconds?: number | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  uncertaintyHandling?: number | null;
  /** @nullable */
  justificationQuality?: number | null;
  /** @nullable */
  sourceAwareness?: number | null;
  /** @nullable */
  regulatoryHallucinationRisk?: number | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  verdict?: string | null;
}

export interface RunResults {
  run: Run;
  summaryByModel: ModelSummary[];
  summaryByTopic: GroupSummary[];
  summaryByDifficulty: GroupSummary[];
  rows: ResultRow[];
}

export interface ArenaOption {
  response: string;
}

export interface ArenaDuel {
  duelToken: string;
  questionId: string;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  difficulty?: string | null;
  optionA: ArenaOption;
  optionB: ArenaOption;
}

export interface ArenaModel {
  provider: string;
  model: string;
}

export type ArenaVoteInputWinner = typeof ArenaVoteInputWinner[keyof typeof ArenaVoteInputWinner];


export const ArenaVoteInputWinner = {
  A: 'A',
  B: 'B',
  tie: 'tie',
} as const;

export interface ArenaVoteInput {
  duelToken: string;
  winner: ArenaVoteInputWinner;
}

export interface ArenaRanking {
  provider: string;
  model: string;
  rating: number;
  games: number;
  wins: number;
  losses: number;
  ties: number;
  /** @nullable */
  winRate?: number | null;
}

export interface ArenaLeaderboard {
  totalVotes: number;
  rankings: ArenaRanking[];
}

export type ArenaVoteResultWinner = typeof ArenaVoteResultWinner[keyof typeof ArenaVoteResultWinner];


export const ArenaVoteResultWinner = {
  A: 'A',
  B: 'B',
  tie: 'tie',
} as const;

export interface ArenaVoteResult {
  modelA: ArenaModel;
  modelB: ArenaModel;
  winner: ArenaVoteResultWinner;
  ratingChangeA: number;
  ratingChangeB: number;
  leaderboard: ArenaLeaderboard;
}