# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: benchmark
    description: Biodiversity LLM benchmark operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /benchmark/config:
    get:
      operationId: getBenchmarkConfig
      tags: [benchmark]
      summary: Benchmark configuration
      description: Available providers, topics, difficulties and dataset metadata.
      responses:
        "200":
          description: Configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BenchmarkConfig"
  /benchmark/runs:
    get:
      operationId: listRuns
      tags: [benchmark]
      summary: List benchmark runs
      responses:
        "200":
          description: List of runs (most recent first)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
    post:
      operationId: createRun
      tags: [benchmark]
      summary: Start a new benchmark run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunInput"
      responses:
        "201":
          description: Run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/admin/session:
    post:
      operationId: verifyAdmin
      tags: [benchmark]
      summary: Verify the admin password (login) for launching/deleting runs
      responses:
        "204":
          description: Password valid
        "401":
          description: Invalid or missing password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Admin actions disabled (no password configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/runs/{runId}:
    get:
      operationId: getRun
      tags: [benchmark]
      summary: Get a single run with live progress
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteRun
      tags: [benchmark]
      summary: Delete a run and its outputs
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/runs/{runId}/results:
    get:
      operationId: getRunResults
      tags: [benchmark]
      summary: Aggregated results for a run
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aggregated results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResults"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/questions:
    get:
      operationId: listQuestions
      tags: [benchmark]
      summary: Preview the benchmark question base
      responses:
        "200":
          description: All questions in the dataset
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionPreview"
  /benchmark/contact:
    post:
      operationId: submitContact
      tags: [benchmark]
      summary: Send a message to Natural Solutions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/arena/duel:
    get:
      operationId: getArenaDuel
      tags: [benchmark]
      summary: Get a blind duel of two model answers to the same question
      description: Pairs two distinct models' already-generated answers to a random question. Identities are hidden until a vote is cast.
      responses:
        "200":
          description: A duel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArenaDuel"
        "404":
          description: Not enough generated answers to build a duel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/arena/vote:
    post:
      operationId: submitArenaVote
      tags: [benchmark]
      summary: Record a blind vote and reveal the two models
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArenaVoteInput"
      responses:
        "201":
          description: Vote recorded and models revealed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArenaVoteResult"
        "400":
          description: Invalid vote or duel token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /benchmark/arena/leaderboard:
    get:
      operationId: getArenaLeaderboard
      tags: [benchmark]
      summary: Cumulative community Elo leaderboard from arena votes
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArenaLeaderboard"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    QuestionPreview:
      type: object
      required: [id, question]
      properties:
        id:
          type: string
        topic:
          type: ["string", "null"]
        subtopic:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
        questionType:
          type: ["string", "null"]
        countryScope:
          type: ["string", "null"]
        question:
          type: string
        expectedAnswerShort:
          type: ["string", "null"]
    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 200
        message:
          type: string
          minLength: 1
          maxLength: 5000
    ContactResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    Provider:
      type: object
      required: [id, defaultModel, available, requiresKey]
      properties:
        id:
          type: string
        defaultModel:
          type: string
        available:
          type: boolean
        requiresKey:
          type: boolean
    BenchmarkConfig:
      type: object
      required:
        - providers
        - topics
        - difficulties
        - questionTypes
        - judgeModel
        - judgeAvailable
        - totalQuestions
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/Provider"
        topics:
          type: array
          items:
            type: string
        difficulties:
          type: array
          items:
            type: string
        questionTypes:
          type: array
          items:
            type: string
        judgeModel:
          type: string
        judgeAvailable:
          type: boolean
        totalQuestions:
          type: integer
    RunInput:
      type: object
      required: [models]
      properties:
        models:
          type: array
          minItems: 1
          items:
            type: string
        limit:
          type: ["integer", "null"]
        topic:
          type: ["string", "null"]
        dryRun:
          type: boolean
          default: false
        noEval:
          type: boolean
          default: false
    Run:
      type: object
      required:
        - id
        - status
        - phase
        - models
        - dryRun
        - noEval
        - completed
        - total
        - createdAt
      properties:
        id:
          type: string
        status:
          type: string
          description: "queued, running, completed, failed or interrupted"
        phase:
          type: string
          description: "empty, query, evaluate, report or done"
        models:
          type: array
          items:
            type: string
        limit:
          type: ["integer", "null"]
        topic:
          type: ["string", "null"]
        dryRun:
          type: boolean
        noEval:
          type: boolean
        completed:
          type: integer
        total:
          type: integer
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
        judgeModel:
          type: ["string", "null"]
    ModelSummary:
      type: object
      required: [provider, model, nQuestions, nErrors]
      properties:
        provider:
          type: string
        model:
          type: string
        size:
          type: ["string", "null"]
          description: "Approximate model size class (e.g. Petit, Moyen, Grand)"
        nQuestions:
          type: integer
        nErrors:
          type: integer
        avgLatency:
          type: ["number", "null"]
        accuracy:
          type: ["number", "null"]
        uncertaintyHandling:
          type: ["number", "null"]
        justificationQuality:
          type: ["number", "null"]
        sourceAwareness:
          type: ["number", "null"]
        regulatoryHallucinationRisk:
          type: ["number", "null"]
        overallScore:
          type: ["number", "null"]
    GroupSummary:
      type: object
      required: [group, provider, model]
      properties:
        group:
          type: string
          description: topic or difficulty label
        provider:
          type: string
        model:
          type: string
        overallScore:
          type: ["number", "null"]
        accuracy:
          type: ["number", "null"]
        uncertaintyHandling:
          type: ["number", "null"]
        justificationQuality:
          type: ["number", "null"]
        sourceAwareness:
          type: ["number", "null"]
        regulatoryHallucinationRisk:
          type: ["number", "null"]
    ResultRow:
      type: object
      required: [questionId, provider, model]
      properties:
        questionId:
          type: string
        topic:
          type: ["string", "null"]
        subtopic:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
        questionType:
          type: ["string", "null"]
        countryScope:
          type: ["string", "null"]
        question:
          type: ["string", "null"]
        provider:
          type: string
        model:
          type: string
        rawResponse:
          type: ["string", "null"]
        latencySeconds:
          type: ["number", "null"]
        error:
          type: ["string", "null"]
        accuracy:
          type: ["number", "null"]
        uncertaintyHandling:
          type: ["number", "null"]
        justificationQuality:
          type: ["number", "null"]
        sourceAwareness:
          type: ["number", "null"]
        regulatoryHallucinationRisk:
          type: ["number", "null"]
        overallScore:
          type: ["number", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        verdict:
          type: ["string", "null"]
    RunResults:
      type: object
      required: [run, summaryByModel, summaryByTopic, summaryByDifficulty, rows]
      properties:
        run:
          $ref: "#/components/schemas/Run"
        summaryByModel:
          type: array
          items:
            $ref: "#/components/schemas/ModelSummary"
        summaryByTopic:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        summaryByDifficulty:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ResultRow"
    ArenaOption:
      type: object
      required: [response]
      properties:
        response:
          type: string
    ArenaDuel:
      type: object
      required: [duelToken, questionId, optionA, optionB]
      properties:
        duelToken:
          type: string
        questionId:
          type: string
        question:
          type: ["string", "null"]
        topic:
          type: ["string", "null"]
        difficulty:
          type: ["string", "null"]
        optionA:
          $ref: "#/components/schemas/ArenaOption"
        optionB:
          $ref: "#/components/schemas/ArenaOption"
    ArenaModel:
      type: object
      required: [provider, model]
      properties:
        provider:
          type: string
        model:
          type: string
    ArenaVoteInput:
      type: object
      required: [duelToken, winner]
      properties:
        duelToken:
          type: string
        winner:
          type: string
          enum: [A, B, tie]
    ArenaRanking:
      type: object
      required: [provider, model, rating, games, wins, losses, ties]
      properties:
        provider:
          type: string
        model:
          type: string
        rating:
          type: number
        games:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        ties:
          type: integer
        winRate:
          type: ["number", "null"]
    ArenaLeaderboard:
      type: object
      required: [totalVotes, rankings]
      properties:
        totalVotes:
          type: integer
        rankings:
          type: array
          items:
            $ref: "#/components/schemas/ArenaRanking"
    ArenaVoteResult:
      type: object
      required: [modelA, modelB, winner, ratingChangeA, ratingChangeB, leaderboard]
      properties:
        modelA:
          $ref: "#/components/schemas/ArenaModel"
        modelB:
          $ref: "#/components/schemas/ArenaModel"
        winner:
          type: string
          enum: [A, B, tie]
        ratingChangeA:
          type: number
        ratingChangeB:
          type: number
        leaderboard:
          $ref: "#/components/schemas/ArenaLeaderboard"