# lib/api-zod/src/generated/types/arenaDuel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArenaOption } from './arenaOption';

export interface ArenaDuel {
  duelToken: string;
  questionId: string;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  difficulty?: string | null;
  optionA: ArenaOption;
  optionB: ArenaOption;
}