# lib/api-zod/src/generated/types/arenaLeaderboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArenaRanking } from './arenaRanking';

export interface ArenaLeaderboard {
  totalVotes: number;
  rankings: ArenaRanking[];
}