# lib/api-zod/src/generated/types/arenaRanking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArenaRanking {
  provider: string;
  model: string;
  rating: number;
  games: number;
  wins: number;
  losses: number;
  ties: number;
  /** @nullable */
  winRate?: number | null;
}