# lib/api-zod/src/generated/types/arenaVoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArenaVoteInputWinner } from './arenaVoteInputWinner';

export interface ArenaVoteInput {
  duelToken: string;
  winner: ArenaVoteInputWinner;
}