# lib/api-zod/src/generated/types/arenaVoteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArenaLeaderboard } from './arenaLeaderboard';
import type { ArenaModel } from './arenaModel';
import type { ArenaVoteResultWinner } from './arenaVoteResultWinner';

export interface ArenaVoteResult {
  modelA: ArenaModel;
  modelB: ArenaModel;
  winner: ArenaVoteResultWinner;
  ratingChangeA: number;
  ratingChangeB: number;
  leaderboard: ArenaLeaderboard;
}